# [quantization] Vectorized quantized relu/relu6