--- a/aten/src/ATen/native/quantized/cpu/qrelu.cpp
+++ b/aten/src/ATen/native/quantized/cpu/qrelu.cpp
@@ -19,21 +19,30 @@ Tensor quantized_relu(const Tensor& qx) {
         qx.q_scale(),
         qx.q_zero_point(),
         qx.suggest_memory_format());
+    using Vec = Vec256<scalar_t>;
     auto iter = TensorIterator::unary_op(qy, qx);
-    cpu_kernel(iter, [&](scalar_t value) -> scalar_t {
-      return scalar_t(std::max<underlying_t>(value.val_, zero_point));
-    });
+    auto zero_point_vec = Vec(scalar_t(zero_point));
+    cpu_kernel_vec(
+        iter,
+        [&](scalar_t value) -> scalar_t {
+          return scalar_t(std::max<underlying_t>(value.val_, zero_point));
+        },
+        [&](Vec value) -> Vec { return value.relu(zero_point_vec); });
   });
   return qy;
 }
-
 Tensor& quantized_relu_(Tensor& qx) {
   const auto zero_point = qx.q_zero_point();
   AT_DISPATCH_QINT_TYPES(qx.scalar_type(), "qrelu", [&]() {
+    using Vec = Vec256<scalar_t>;
     auto iter = TensorIterator::unary_op(qx, qx);
-    cpu_kernel(iter, [&](scalar_t value) -> scalar_t {
-      return scalar_t(std::max<underlying_t>(value.val_, zero_point));
-    });
+    auto zero_point_vec = Vec(scalar_t(zero_point));
+    cpu_kernel_vec(
+        iter,
+        [&](scalar_t value) -> scalar_t {
+          return scalar_t(std::max<underlying_t>(value.val_, zero_point));
+        },
+        [&](Vec value) -> Vec { return value.relu(zero_point_vec); });
   });
   return qx;
 }
@@ -49,13 +58,20 @@ Tensor quantized_relu6(const Tensor& qx) {
         qx.q_scale(),
         qx.q_zero_point(),
         qx.suggest_memory_format());
+    using Vec = Vec256<scalar_t>;
     auto iter = TensorIterator::unary_op(qy, qx);
     scalar_t six = at::quantize_val<scalar_t>(qx.q_scale(), qx.q_zero_point(),
                                               6.0);
-    cpu_kernel(iter, [&](scalar_t value) -> scalar_t {
-      underlying_t relu_val = std::max<underlying_t>(value.val_, zero_point);
-      return scalar_t(std::min<underlying_t>(relu_val, six.val_));
-    });
+    auto zero_point_vec = Vec(scalar_t(zero_point));
+    auto six_vec = Vec(six);
+    cpu_kernel_vec(
+        iter,
+        [&](scalar_t value) -> scalar_t {
+          underlying_t relu_val =
+              std::max<underlying_t>(value.val_, zero_point);
+          return scalar_t(std::min<underlying_t>(relu_val, six.val_));
+        },
+        [&](Vec val) { return val.relu6(zero_point_vec, six_vec); });
   });
   return qy;
 }